+++ src/sll/gc.c
@@ -5,7 +5,6 @@
 #include <sll/array.h>
 #include <sll/common.h>
 #include <sll/data.h>
-#include <sll/file.h>
 #include <sll/gc.h>
 #include <sll/map.h>
 #include <sll/memory.h>